refactory: Corrige o namespace sos controllers cujas pastas foram postas no plural. files: - app/Controllers/Admins/BathroomsController.php - app/Controllers/Admins/BuildingController.php - app/Controllers/Admins/HomeController.php - app/Controllers/Users/HomeController.php
