+++ app/Controllers/Users/HomeController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controllers\User;
+namespace App\Controllers\Users;
 
 use Core\Http\Controllers\Controller;
 use Core\Http\Request;
